+++ addOns/reports/src/main/java/org/zaproxy/addon/reports/ReportDialog.java
@@ -330,6 +330,7 @@ public class ReportDialog extends StandardFieldsDialog {
             }
         }
         getSectionsScrollPane().setViewportView(sectionPanel);
+        repaint();
     }
 
     private void setReportName() {