+++ addOns/reports/CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Fixed
-- Fixed Sections Option not fully shown in Generate Report Dialog.
+- Ensure the Sections' options are fully shown always in Generate Report Dialog (Issue 8259).
 
 ## [0.27.0] - 2023-12-19
 ### Changed